+++ lua/DCS-gRPC/methods/unit.lua
@@ -66,25 +66,9 @@ GRPC.methods.getUnitTransform = function(params)
     return GRPC.errorNotFound("unit does not exist")
   end
 
-    -- https://wiki.hoggitworld.com/view/DCS_func_getPosition
-  local position = unit:getPosition()
-  local coords = GRPC.exporters.position(position.p)
-  local coordsNorth = coord.LLtoLO(coords.lat + 1, coords.lon)
-
-  -- Response does not exactly match what the gRPC server will return since the gRPC response
-  -- will be calculated based on the information returned here.
   return GRPC.success({
-    position = coords,
-    u = position.p.z,
-    v = position.p.x,
-    positionNorth = GRPC.exporters.vector(coordsNorth),
-    orientation = {
-      forward = GRPC.exporters.vector(position.x),
-      right = GRPC.exporters.vector(position.z),
-      up = GRPC.exporters.vector(position.y),
-    },
-    velocity = GRPC.exporters.vector(unit:getVelocity()),
     time = timer.getTime(),
+    rawTransform = GRPC.exporters.rawTransform(unit),
   })
 end
 
@@ -133,18 +117,4 @@ GRPC.methods.getUnit = function(params)
   end
 
   return GRPC.success({unit = GRPC.exporters.unit(unit)})
-end
-
--- This method is only used by the unit stream. It returns the subset of `getUnit` that can
--- change (like position, velocity, ...). It is not exposed to the gRPC service for now.
-GRPC.methods.getUnitUpdate = function(params)
-  local unit = Unit.getByName(params.name)
-  if unit == nil then
-    return GRPC.errorNotFound("unit `" .. tostring(params.name) .. "` does not exist")
-  end
-
-  return GRPC.success({
-    position = GRPC.exporters.position(unit:getPoint()),
-    velocity = GRPC.exporters.vector(unit:getVelocity()),
-  })
 end