+++ stubs/src/custom.rs
@@ -1,0 +1,3 @@
+pub mod v0 {
+    tonic::include_proto!("dcs.custom.v0");
+}